Add filter routine to exclude RTs and replies
By default, the Twitter Streaming API .filter method will include not
just original tweets, but replies and retweets to the filtered user as
well. This is unintended behavior for this bot, so a routine to exclude
these results is necessary.

The routine currently checks each tweet's user id against the array of
ids stored in the bot's configuration. This works, but is not perfect,
as it has the potential, when monitoring multiple ids, to return replies
or RTs between those ids.